owrt-qti-conf: Add logic to allow for execution of build.py from any workspace directory
This commit adds the get_owrt_root_path() function to the build.py script.
The function is responsible for determining the path to the OpenWrt
build system's root directory (owrt/). It uses the current script's
directory as a starting point and navigates up one level to find the
parent directory. From there, it navigates up again to reach the OpenWrt
root directory. The get_owrt_root_path() function returns the computed path
as a string, which is then assigned to TOPDIR variable, providing a reliable
way to locate the root directory of the OpenWrt build system and execute the
build script regardless of the current direcotry the user is operating on.

Change-Id: Iffdd0d1d107963422de9ebcc52e15bb605b2806c